requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.12
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.54.0
httpx==0.27.2
greenlet>=3.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
streamlit==1.39.0
requests==2.32.3
